# Optimize String#scan and make MatchData ruby/spec compliant