--- a/mruby-bin/ruby/Gemfile
+++ b/mruby-bin/ruby/Gemfile
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+source 'https://rubygems.org'
+
+gem 'rubocop'