--- a/mruby/src/convert/hash.rs
+++ b/mruby/src/convert/hash.rs
@@ -255,12 +255,12 @@ impl FromMrb<Vec<(&str, Value)>> for Value {
     }
 }
 
-impl FromMrb<HashMap<&str, Value>> for Value {
+impl FromMrb<HashMap<&str, Self>> for Value {
     type From = Rust;
     type To = Ruby;
 
-    fn from_mrb(interp: &Mrb, value: HashMap<&str, Value>) -> Self {
-        let pairs = value.into_iter().collect::<Vec<(&str, Value)>>();
+    fn from_mrb(interp: &Mrb, value: HashMap<&str, Self>) -> Self {
+        let pairs = value.into_iter().collect::<Vec<(&str, Self)>>();
         Self::from_mrb(interp, pairs)
     }
 }