--- a/mruby/src/extn/core/matchdata/matchdata.rb
+++ b/mruby/src/extn/core/matchdata/matchdata.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class MatchData
+  def values_at(*indexes)
+    indexes.map { |index| self[index] }.flatten
+  end
+end