--- a/mruby/src/extn/core/matchdata/matchdata.rb
+++ b/mruby/src/extn/core/matchdata/matchdata.rb
@@ -1,6 +1,33 @@
 # frozen_string_literal: true
 
 class MatchData
+  def ==(other)
+    return false unless other.is_a?(MatchData)
+    return false unless string == other.string
+    return false unless regexp == other.regexp
+    return false unless offset(0) == other.offset(0)
+
+    true
+  end
+
+  def eql?(other)
+    self == other
+  end
+
+  def inspect
+    s = %(#<MatchData "#{self[0]}")
+    if named_captures.empty?
+      captures.each_with_index do |capture, index|
+        s << %( #{index + 1}:"#{capture || nil.inspect}")
+      end
+    else
+      named_captures.each do |(group, capture)|
+        s << %( #{group}:"#{capture || nil.inspect}")
+      end
+    end
+    s << '>'
+  end
+
   def values_at(*indexes)
     indexes.map { |index| self[index] }.flatten
   end