--- a/scripts/lint.sh
+++ b/scripts/lint.sh
@@ -25,6 +25,8 @@ cargo doc --no-deps --all
 lint_ruby_sources mruby/src/extn
 ## spec-runner
 lint_ruby_sources spec-runner/src
+## mruby bins
+lint_ruby_sources mruby-bin/ruby
 ## nemesis
 lint_ruby_sources nemesis/ruby
 ## foolsgold