`String#scan` was implemented in Ruby and was slow and not spec compliant. This PR makes `String#scan` ruby/spec compliant and improves performance by a factor of 20. The implementation of `String#scan` in this crate is roughly comparable to CRuby for normal workloads (slower than CRuby by about 10-20%), but can be up to 50% faster for large Strings with sparse matches.

To do that, implement `String#scan` in Rust using an onig `Regex` directly. Before this PR, `String#scan` spends a lot of time in mruby `String#[]`.

<img width="1189" alt="Screen Shot 2019-06-27 at 5 12 40 PM" src="https://user-images.githubusercontent.com/860434/60561038-f8bcd400-9d06-11e9-90e0-d2d8b84c205b.png">

It now spends all of its time in onig.

<img width="1390" alt="Screen Shot 2019-07-02 at 8 20 12 PM" src="https://user-images.githubusercontent.com/860434/60561045-fe1a1e80-9d06-11e9-819e-aa1a13b7dd13.png">

This PR adds a benchmark to mruby-bin which is runnable like this:

```console
$ cargo run --release --bin string_scan_bench -- mruby-bin/ruby/benches/string_scan.rb
String#scan bench for mruby 2.0 [2.0.1-11]

Email: 92 matches
..................................................

    compile: 5.47ms elapsed in 50 iterations (avg. 0.1ms / iteration)
    scan: 16570.2ms elapsed in 50 iterations (avg. 331.4ms / iteration)
    scan with block: 16510.32ms elapsed in 50 iterations (avg. 330.2ms / iteration)

URI: 5301 matches
..................................................

    compile: 2.8ms elapsed in 50 iterations (avg. 0.05ms / iteration)
    scan: 15635.23ms elapsed in 50 iterations (avg. 312.7ms / iteration)
    scan with block: 15533.86ms elapsed in 50 iterations (avg. 310.67ms / iteration)

IP: 5 matches
..................................................

    compile: 2.33ms elapsed in 50 iterations (avg. 0.04ms / iteration)
    scan: 1386.49ms elapsed in 50 iterations (avg. 27.72ms / iteration)
    scan with block: 1367.78ms elapsed in 50 iterations (avg. 27.35ms / iteration)
```

and for CRuby like this:

```console
$ ruby mruby-bin/ruby/benches/string_scan.rb
String#scan bench for ruby 2.6.3p62 (2019-04-16 revision 67580) [x86_64-darwin18]

Email: 92 matches
..................................................

    compile: 1.84ms elapsed in 50 iterations (avg. 0.03ms / iteration)
    scan: 14724.34ms elapsed in 50 iterations (avg. 294.48ms / iteration)
    scan with block: 14669.76ms elapsed in 50 iterations (avg. 293.39ms / iteration)

URI: 5301 matches
..................................................

    compile: 1.93ms elapsed in 50 iterations (avg. 0.03ms / iteration)
    scan: 12823.31ms elapsed in 50 iterations (avg. 256.46ms / iteration)
    scan with block: 12786.91ms elapsed in 50 iterations (avg. 255.73ms / iteration)

IP: 5 matches
..................................................

    compile: 2.32ms elapsed in 50 iterations (avg. 0.04ms / iteration)
    scan: 2714.96ms elapsed in 50 iterations (avg. 54.29ms / iteration)
    scan with block: 2661.23ms elapsed in 50 iterations (avg. 53.22ms / iteration)
```

To pass ruby/spec for `String#scan`, `MatchData` needed to support windowed matches on a source `String`. I took the opportunity to refactor `MatchData`'s Rust representation to hold a match `Region`. I also implemented all of the missing methods   and fixed up some spec compliance issues in the existing implementation. Apart from the `` $` `` and `$'` special variables, `MatchData` is fully compliant with ruby/spec.

Fixes GH-135.